+++ example/ViewController.m
@@ -60,7 +60,7 @@
 #pragma mark - RNFrostedSidebarDelegate
 
 - (void)sidebar:(RNFrostedSidebar *)sidebar didTapItemAtIndex:(NSUInteger)index {
-    NSLog(@"Tapped item at index %i",index);
+    NSLog(@"Tapped item at index %lu",(unsigned long)index);
     if (index == 3) {
         [sidebar dismissAnimated:YES completion:nil];
     }
